Re-org of commands
